Moved httpClientHelper to Helpers directory
